requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.0
google-generativeai==0.3.1
pyyaml==6.0.1
loguru==0.7.0 